% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_constant.R
\name{createDist_Const}
\alias{createDist_Const}
\title{Create Constant Distribution}
\usage{
createDist_Const(box)
}
\value{
a distribution object of a constant distribution
}
\description{
Create Constant Distribution
}
