% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_exponential.R
\name{createDist_exp}
\alias{createDist_exp}
\title{Create Multivariate Exponential Distribution}
\usage{
createDist_exp(coefs, box, prior = list())
}
\value{
a distribution object of an exponential distribution
}
\description{
Create Multivariate Exponential Distribution
}
