% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mixture.R
\name{createDist_Mix}
\alias{createDist_Mix}
\title{Create mixture distribution}
\usage{
createDist_Mix(prop = numeric(0), compList = list(), prior = list())
}
\arguments{
\item{prop}{vector of probabilities for the components}

\item{compList}{list of distributions}
}
\value{
a distribution object containing a mixture distribution
}
\description{
Create mixture distribution
}
