% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_mvn.R
\name{createDist_MVN}
\alias{createDist_MVN}
\title{Create Multivariate Normal Distribution}
\usage{
createDist_MVN(mean, sigma, prior = list())
}
\arguments{
\item{mean}{the center vector}

\item{sigma}{the covariance matrix}
}
\value{
a distribution object of a multivariate normal distribution
}
\description{
Create Multivariate Normal Distribution
}
