% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getDensity}
\alias{getDensity}
\title{Get density}
\usage{
getDensity(dist, x, log = TRUE, vars = NULL)
}
\arguments{
\item{dist}{a distribution object}

\item{x}{contains \code{p x n} matrix where columns are observation vectors}

\item{log}{if TRUE return logarithm of density}
}
\value{
a vector with the (log) densities for the observation vectors
}
\description{
Get density
}
