% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getMembership}
\alias{getMembership}
\title{Get membership probability}
\usage{
getMembership(dist, x, log = TRUE)
}
\arguments{
\item{dist}{a distribution object}

\item{x}{contains \code{p x n} matrix where columns are observation vectors}

\item{log}{if TRUE returns logarithm of membership probabilities}
}
\value{
a \code{m x n} matrix where each column is associated with an observation vector
        and contains for each of the \code{m} components the probability of membership
}
\description{
Get membership probability
}
\note{
Only distributions which are mixtures of distributions implement this function.
}
