% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getMLEstimate}
\alias{getMLEstimate}
\title{Maximum Likelihood Estimate}
\usage{
getMLEstimate(dist, x, weights = rep(1, ncol(x)), ...)
}
\arguments{
\item{dist}{a distribution object}

\item{x}{contains \code{p x n} matrix where columns are observation vectors}

\item{weights}{importance of each observation}

\item{...}{distribution specific tuning parameters}
}
\value{
a distribution object with optimized distribution parameters
}
\description{
Maximum Likelihood Estimate
}
