% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getPosteriorSample}
\alias{getPosteriorSample}
\title{Get Posterior Sample}
\usage{
getPosteriorSample(dist, x, n, ...)
}
\arguments{
\item{dist}{a distribution object}

\item{x}{contains \code{p x n} matrix where columns are observation vectors}

\item{n}{number of samples}

\item{...}{distribution specific tuning parameters}
}
\value{
a sample from the posterior distribution (via Gibbs sampling)
}
\description{
Get Posterior Sample
}
