% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getPriorDensity}
\alias{getPriorDensity}
\title{Get Prior Density}
\usage{
getPriorDensity(dist, ...)
}
\arguments{
\item{dist}{a distribution object}

\item{...}{distribution specific tuning parameters}

\item{n}{number of samples}
}
\description{
Get Prior Density
}
