% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_generic.R
\name{getSample}
\alias{getSample}
\title{Draw a sample}
\usage{
getSample(dist, n)
}
\arguments{
\item{dist}{a distribution object}

\item{n}{number of samples to draw from distribution}
}
\value{
a \code{p x n} matrix, each column contains an observation vector
}
\description{
Draw a sample
}
